backend/pyproject.toml
[project]
name = "autobay"
version = "0.1.0"
description = "AI-powered cross-border e-commerce automation platform"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",

    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "greenlet>=3.1.0",

    # Redis
    "redis>=5.2.0",

    # Task queue
    "celery[redis]>=5.4.0",

    # Auth
    "pyjwt>=2.10.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.2.0",

    # Config
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "email-validator>=2.2.0",

    # HTTP client (for platform integrations)
    "httpx>=0.28.0",

    # AI/LLM
    "anthropic>=0.40.0",
    "openai>=1.55.0",

    # Utilities
    "python-dateutil>=2.9.0",
    "orjson>=3.10.0",

    # Resilience
    "tenacity>=9.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "factory-boy>=3.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
